// apps/web/src/shared/api/graphql/__generated__/schema.graphql.ts
export type Maybe<T> = T | null;
// ...
export interface Scalars {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: any; output: any };
  DateTime: { input: any; output: any };
  JSON: { input: any; output: any };
  Time: { input: any; output: any };
  Upload: { input: any; output: any };
}

export interface About {
  __typename?: 'About';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  textEn?: Maybe<Scalars['String']['output']>;
  textRu?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface AboutEntity {
  __typename?: 'AboutEntity';
  attributes?: Maybe<About>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface AboutEntityResponse {
  __typename?: 'AboutEntityResponse';
  data?: Maybe<AboutEntity>;
}
// ...
export interface ComponentLocationLocation {
  __typename?: 'ComponentLocationLocation';
  id: Scalars['ID']['output'];
  locationLink?: Maybe<Scalars['String']['output']>;
  locationName?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Cover {
  __typename?: 'Cover';
  url: Scalars['String']['output'];
}

export interface CustomDate {
  __typename?: 'CustomDate';
  day: Scalars['String']['output'];
  week: Scalars['String']['output'];
}
// ...
export interface EventEntity {
  __typename?: 'EventEntity';
  attributes?: Maybe<Event>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface EventEntityResponse {
  __typename?: 'EventEntityResponse';
  data?: Maybe<EventEntity>;
}

export interface EventEntityResponseCollection {
  __typename?: 'EventEntityResponseCollection';
  data: Array<EventEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface EventScheduleEntity {
  __typename?: 'EventScheduleEntity';
  attributes?: Maybe<EventSchedule>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface EventScheduleEntityResponse {
  __typename?: 'EventScheduleEntityResponse';
  data?: Maybe<EventScheduleEntity>;
}

export interface EventScheduleEntityResponseCollection {
  __typename?: 'EventScheduleEntityResponseCollection';
  data: Array<EventScheduleEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface Genre {
  __typename?: 'Genre';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  mixes?: Maybe<MixRelationResponseCollection>;
  name: Scalars['String']['output'];
  slug: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface GenreMixesArgs {
  filters?: InputMaybe<MixFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}

export interface GenreEntity {
  __typename?: 'GenreEntity';
  attributes?: Maybe<Genre>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface GenreEntityResponse {
  __typename?: 'GenreEntityResponse';
  data?: Maybe<GenreEntity>;
}

export interface GenreEntityResponseCollection {
  __typename?: 'GenreEntityResponseCollection';
  data: Array<GenreEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface GuestEntity {
  __typename?: 'GuestEntity';
  attributes?: Maybe<Guest>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface GuestEntityResponse {
  __typename?: 'GuestEntityResponse';
  data?: Maybe<GuestEntity>;
}

export interface GuestEntityResponseCollection {
  __typename?: 'GuestEntityResponseCollection';
  data: Array<GuestEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface I18NLocaleEntity {
  __typename?: 'I18NLocaleEntity';
  attributes?: Maybe<I18NLocale>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface I18NLocaleEntityResponse {
  __typename?: 'I18NLocaleEntityResponse';
  data?: Maybe<I18NLocaleEntity>;
}

export interface I18NLocaleEntityResponseCollection {
  __typename?: 'I18NLocaleEntityResponseCollection';
  data: Array<I18NLocaleEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface Mix {
  __typename?: 'Mix';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  date: Scalars['DateTime']['output'];
  descriptionEn?: Maybe<Scalars['String']['output']>;
  descriptionRu?: Maybe<Scalars['String']['output']>;
  event?: Maybe<EventEntityResponse>;
  genres?: Maybe<GenreRelationResponseCollection>;
  guests?: Maybe<GuestRelationResponseCollection>;
  image: UploadFileEntityResponse;
  linksToMixes?: Maybe<ComponentLinksToMixesLink>;
  locationLink?: Maybe<Scalars['String']['output']>;
  locationName?: Maybe<Scalars['String']['output']>;
  moods?: Maybe<MoodRelationResponseCollection>;
  name: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  show?: Maybe<ShowEntityResponse>;
  slug?: Maybe<Scalars['String']['output']>;
  tracklist?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}
// ...
export interface MixEntity {
  __typename?: 'MixEntity';
  attributes?: Maybe<Mix>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface MixEntityResponse {
  __typename?: 'MixEntityResponse';
  data?: Maybe<MixEntity>;
}

export interface MixEntityResponseCollection {
  __typename?: 'MixEntityResponseCollection';
  data: Array<MixEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface MixRelationResponseCollection {
  __typename?: 'MixRelationResponseCollection';
  data: Array<MixEntity>;
}

export interface Mood {
  __typename?: 'Mood';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  mixes?: Maybe<MixRelationResponseCollection>;
  name?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
}

export interface MoodMixesArgs {
  filters?: InputMaybe<MixFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
}

export interface MoodEntity {
  __typename?: 'MoodEntity';
  attributes?: Maybe<Mood>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface MoodEntityResponse {
  __typename?: 'MoodEntityResponse';
  data?: Maybe<MoodEntity>;
}

export interface MoodEntityResponseCollection {
  __typename?: 'MoodEntityResponseCollection';
  data: Array<MoodEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface Pagination {
  __typename?: 'Pagination';
  page: Scalars['Int']['output'];
  pageCount: Scalars['Int']['output'];
  pageSize: Scalars['Int']['output'];
  total: Scalars['Int']['output'];
}

export interface PaginationArg {
  limit?: InputMaybe<Scalars['Int']['input']>;
  page?: InputMaybe<Scalars['Int']['input']>;
  pageSize?: InputMaybe<Scalars['Int']['input']>;
  start?: InputMaybe<Scalars['Int']['input']>;
}

export interface PopularityResponse {
  __typename?: 'PopularityResponse';
  cover?: Maybe<Cover>;
  date: Scalars['DateTime']['output'];
  fixedDate?: Maybe<CustomDate>;
  id: Scalars['ID']['output'];
  location?: Maybe<ComponentLocationLocation>;
  schedule?: Maybe<Array<Maybe<Schedule>>>;
}

export enum PublicationState {
  Live = 'LIVE',
  Preview = 'PREVIEW',
}

export interface Query {
  __typename?: 'Query';
  about?: Maybe<AboutEntityResponse>;
  event?: Maybe<EventEntityResponse>;
  eventSchedule?: Maybe<EventScheduleEntityResponse>;
  eventSchedules?: Maybe<EventScheduleEntityResponseCollection>;
  eventSchedulesFixed?: Maybe<Array<Maybe<PopularityResponse>>>;
  events?: Maybe<EventEntityResponseCollection>;
  genre?: Maybe<GenreEntityResponse>;
  genreOne?: Maybe<GenreEntityResponse>;
  genres?: Maybe<GenreEntityResponseCollection>;
  genresArray?: Maybe<Array<Maybe<Array<Maybe<Genre>>>>>;
  guest?: Maybe<GuestEntityResponse>;
  guests?: Maybe<GuestEntityResponseCollection>;
  homePageRandomMix?: Maybe<MixEntityResponse>;
  i18NLocale?: Maybe<I18NLocaleEntityResponse>;
  i18NLocales?: Maybe<I18NLocaleEntityResponseCollection>;
  me?: Maybe<UsersPermissionsMe>;
  mix?: Maybe<MixEntityResponse>;
  mixes?: Maybe<MixEntityResponseCollection>;
  mood?: Maybe<MoodEntityResponse>;
  moods?: Maybe<MoodEntityResponseCollection>;
  moodsArray?: Maybe<Array<Maybe<Array<Maybe<Mood>>>>>;
  randomMix?: Maybe<MixEntityResponse>;
  randomMixes?: Maybe<MixEntityResponseCollection>;
  release?: Maybe<ReleaseEntityResponse>;
  releases?: Maybe<ReleaseEntityResponseCollection>;
  shopCategories?: Maybe<ShopCategoryEntityResponseCollection>;
  shopCategory?: Maybe<ShopCategoryEntityResponse>;
  shopItem?: Maybe<ShopItemEntityResponse>;
  shopItems?: Maybe<ShopItemEntityResponseCollection>;
  shopOrder?: Maybe<ShopOrderEntityResponse>;
  shopOrders?: Maybe<ShopOrderEntityResponseCollection>;
  show?: Maybe<ShowEntityResponse>;
  shows?: Maybe<ShowEntityResponseCollection>;
  streamIsLive?: Maybe<Scalars['Boolean']['output']>;
  uploadFile?: Maybe<UploadFileEntityResponse>;
  uploadFiles?: Maybe<UploadFileEntityResponseCollection>;
  uploadFolder?: Maybe<UploadFolderEntityResponse>;
  uploadFolders?: Maybe<UploadFolderEntityResponseCollection>;
  usersPermissionsRole?: Maybe<UsersPermissionsRoleEntityResponse>;
  usersPermissionsRoles?: Maybe<UsersPermissionsRoleEntityResponseCollection>;
  usersPermissionsUser?: Maybe<UsersPermissionsUserEntityResponse>;
  usersPermissionsUsers?: Maybe<UsersPermissionsUserEntityResponseCollection>;
  youtubeLink?: Maybe<YoutubeLinkEntityResponse>;
}
// ...
export interface ReleaseEntityResponse {
  __typename?: 'ReleaseEntityResponse';
  data?: Maybe<ReleaseEntity>;
}

export interface ReleaseEntityResponseCollection {
  __typename?: 'ReleaseEntityResponseCollection';
  data: Array<ReleaseEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface ResponseCollectionMeta {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
}

export interface Schedule {
  __typename?: 'Schedule';
  artist?: Maybe<Artist>;
  time?: Maybe<Scalars['String']['output']>;
}
// ...
export interface ShopCategoryEntityResponse {
  __typename?: 'ShopCategoryEntityResponse';
  data?: Maybe<ShopCategoryEntity>;
}

export interface ShopCategoryEntityResponseCollection {
  __typename?: 'ShopCategoryEntityResponseCollection';
  data: Array<ShopCategoryEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface ShopItemEntityResponse {
  __typename?: 'ShopItemEntityResponse';
  data?: Maybe<ShopItemEntity>;
}

export interface ShopItemEntityResponseCollection {
  __typename?: 'ShopItemEntityResponseCollection';
  data: Array<ShopItemEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface ShopOrderEntityResponse {
  __typename?: 'ShopOrderEntityResponse';
  data?: Maybe<ShopOrderEntity>;
}

export interface ShopOrderEntityResponseCollection {
  __typename?: 'ShopOrderEntityResponseCollection';
  data: Array<ShopOrderEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface ShowEntity {
  __typename?: 'ShowEntity';
  attributes?: Maybe<Show>;
  id?: Maybe<Scalars['ID']['output']>;
}

export interface ShowEntityResponse {
  __typename?: 'ShowEntityResponse';
  data?: Maybe<ShowEntity>;
}

export interface ShowEntityResponseCollection {
  __typename?: 'ShowEntityResponseCollection';
  data: Array<ShowEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface UploadFileEntityResponse {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
}

export interface UploadFileEntityResponseCollection {
  __typename?: 'UploadFileEntityResponseCollection';
  data: Array<UploadFileEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface UploadFolderEntityResponse {
  __typename?: 'UploadFolderEntityResponse';
  data?: Maybe<UploadFolderEntity>;
}

export interface UploadFolderEntityResponseCollection {
  __typename?: 'UploadFolderEntityResponseCollection';
  data: Array<UploadFolderEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface UsersPermissionsMe {
  __typename?: 'UsersPermissionsMe';
  blocked?: Maybe<Scalars['Boolean']['output']>;
  confirmed?: Maybe<Scalars['Boolean']['output']>;
  email?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  role?: Maybe<UsersPermissionsMeRole>;
  username: Scalars['String']['output'];
}
// ...
export interface UsersPermissionsRoleEntityResponse {
  __typename?: 'UsersPermissionsRoleEntityResponse';
  data?: Maybe<UsersPermissionsRoleEntity>;
}

export interface UsersPermissionsRoleEntityResponseCollection {
  __typename?: 'UsersPermissionsRoleEntityResponseCollection';
  data: Array<UsersPermissionsRoleEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface UsersPermissionsUserEntityResponse {
  __typename?: 'UsersPermissionsUserEntityResponse';
  data?: Maybe<UsersPermissionsUserEntity>;
}

export interface UsersPermissionsUserEntityResponseCollection {
  __typename?: 'UsersPermissionsUserEntityResponseCollection';
  data: Array<UsersPermissionsUserEntity>;
  meta: ResponseCollectionMeta;
}
// ...
export interface YoutubeLinkEntityResponse {
  __typename?: 'YoutubeLinkEntityResponse';
  data?: Maybe<YoutubeLinkEntity>;
}